added more docs to the code and fixed the example config
