+++ yaca-voice/common/index.ts
@@ -1,3 +1,8 @@
+/**
+ * Sleeps for a given amount of time.
+ *
+ * @param ms - The amount of time to sleep in milliseconds.
+ */
 export function sleep(ms: number) {
     return new Promise((resolve) => setTimeout(resolve, ms, null));
 }
