+++ yaca-voice/common/locale/index.ts
@@ -3,6 +3,13 @@ import { printf } from 'fast-printf';
 const resourceName = GetCurrentResourceName();
 const dict: Record<string, string> = {};
 
+/**
+ * Flattens a dictionary.
+ *
+ * @param source - The source dictionary to flatten.
+ * @param target - The target dictionary to flatten to.
+ * @param prefix - The prefix to use.
+ */
 function flattenDict(source: Record<string, string | number | boolean>, target: Record<string, string>, prefix?: string) {
     for (const [key, value] of Object.entries(source)) {
         const fullKey = prefix ? `${prefix}.${key}` : key;
@@ -14,6 +21,12 @@ function flattenDict(source: Record<string, string | number | boolean>, target: 
     return target;
 }
 
+/**
+ * Get the localized string for a key.
+ *
+ * @param str - The key to get the localized string for.
+ * @param args - The arguments to use for string interpolation.
+ */
 export const locale = (str: string, ...args: (string | number | boolean)[]): string => {
     const localeStr = dict[str];
 
@@ -28,8 +41,16 @@ export const locale = (str: string, ...args: (string | number | boolean)[]): str
     }
 };
 
+/**
+ * Get all the locales.
+ */
 export const getLocales = () => dict;
 
+/**
+ * Initialize the locale.
+ *
+ * @param configLocale - The locale to use. Defaults to 'en'. If not found, falls back to 'en'.
+ */
 export const initLocale = (configLocale: string) => {
     const lang = configLocale || 'en'
     let locales: typeof dict = JSON.parse(LoadResourceFile(resourceName, `locales/${lang}.json`));
