+++ yaca-voice/config/shared.json.example
@@ -14,7 +14,7 @@
     "megaphone": "B"
   },
   "maxRadioChannels": 6,
-  shortRadioRange: 150,
+  "shortRadioRange": 150,
   "voiceRange": {
     "defaultIndex": 2,
     "ranges": [1, 3, 8, 15, 20, 25, 30, 40]
