+++ yaca-voice/eslint.config.js
@@ -2,12 +2,10 @@
 
 import eslint from "@eslint/js";
 import tsEslint from "typescript-eslint";
-// import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
 
 export default tsEslint.config(
   eslint.configs.recommended,
   ...tsEslint.configs.strict,
-  // eslintPluginPrettierRecommended,
   {
     rules: {
       "eqeqeq": "error",
