+++ yaca-voice/server/utils/cache.ts
@@ -1,5 +1,8 @@
 import { ServerCache } from "types";
 
+/**
+ * Cached values for the server.
+ */
 export const cache: ServerCache = {
   resource: GetCurrentResourceName(),
 };
