+++ yaca-voice/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./enums";
 export * from "./types";
+export * from "./config";
